rec/server/requirements.txt
# Serafis Evaluation Framework Server Requirements

# Web framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# API clients
openai>=1.10.0
google-generativeai>=0.3.0

# Data processing
numpy>=1.24.0

# HTTP client (for test runner)
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Optional: Development
# pytest>=8.0.0
# httpx>=0.26.0  # For async testing
